% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TFBS_similarity}
\alias{TFBS_similarity}
\title{Information on the similarity of transcription factor binding sites on the database TFDB}
\format{
A data frame with 191271 rows and 7 variables:
\describe{
  \item{PCC}{PCC}
  \item{PCC_normalized}{PCC normalized}
  \item{gene1}{gene 1}
  \item{gene2}{gene 2}
}
}
\usage{
TFBS_similarity
}
\description{
Information on the similarity of transcription factor binding sites on the database TFDB
}
\keyword{datasets}
