% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sim_quantile.R
\name{get_sim_quantile}
\alias{get_sim_quantile}
\title{Function to}
\usage{
get_sim_quantile(x = 0.99, distance_measure = "PCC", normalized = TRUE)
}
\arguments{
\item{distance_measure}{PCC or KL}

\item{normalized}{Whether normalized similarity should be used}
}
\value{
Numeric.
}
\description{
Function to
}
